backend/agents/pdf_translator_v2/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pymupdf==1.24.5
anthropic==0.40.0
python-dotenv==1.0.1
langgraph==0.2.28
langchain-core==0.3.0
